fix: uses int64 helper for activity id's

+++ projects/api/src/contracts/checkin/_internal/response/showCheckinResponseSchema.ts
@@ -1,9 +1,9 @@
 import { episodeIdsResponseSchema } from '../../../_internal/response/episodeIdsResponseSchema.ts';
 import { showIdsResponseSchema } from '../../../_internal/response/showIdsResponseSchema.ts';
-import { z } from '../../../_internal/z.ts';
+import { int64, z } from '../../../_internal/z.ts';
 
 export const showCheckinResponseSchema = z.object({
-  id: z.number().int(),
+  id: int64(z.number().int()),
   watched_at: z.string().datetime(),
   sharing: z.object({
     twitter: z.boolean(),
