+++ projects/api/src/contracts/sync/_internal/request/historyIdsRequestSchema.ts
@@ -1,5 +1,5 @@
-import { z } from '../../../_internal/z.ts';
+import { int64, z } from '../../../_internal/z.ts';
 
 export const historyIdsRequestSchema = z.object({
-  ids: z.array(z.number().int()).nullish(),
+  ids: z.array(int64(z.number().int())).nullish(),
 });
